KERI-2: Windows-specific Makefile
This Makefile does what I want the way I want it to.
This Makefile is intended to be "include"d in the original Makefile but
it also works as an explicit stand-alone (if you know what you're
doing).
Tested locally in Windows 10 with Powershell

$PSVersionTable

Name                           Value
----                           -----
PSVersion                      5.1.22621.963
PSEdition                      Desktop
PSCompatibleVersions           {1.0, 2.0, 3.0, 4.0...}
BuildVersion                   10.0.22621.963
CLRVersion                     4.0.30319.42000
WSManStackVersion              3.0
PSRemotingProtocolVersion      2.3
SerializationVersion           1.1.0.1

Changes to be committed:
	new file:   Makefile_windows

Changes not staged for commit:
	modified:   Makefile