KERI-2: Refactored(?) the Makefile
I extricated the necessary functionality into the Makefile_windows and
used this "default" Makefile as an OS-specific level of indirection.
It includes the windows Makefile when running in Windows.

Next, I need to clean up the refactor and then make a Linux-specific
Makefile mirroring the Windows work.

Changes to be committed:
	modified:   Makefile

Untracked files:
	Makefile_linux