KERI-2: Linux-specific Makefile
This is the Makefile the default Makefile will include on a non-Windows
system.  It was copy/pasted from the main, some inclusions from the
Windows makefile copy/pasted in, utilizes the constants makefile and
should be mostly-functional when tested on an actual Linux system.

Next steps:
[ ] Test this file on a Linux system (and fix it)
[ ] Commit and push
[ ] Verify default functionality, as-is, on both Windows and Linux
[ ] Declare good-happy-mission-success

Changes to be committed:
	new file:   Makefile_linux