KERI-2: Updated README w/ Windows OS "gotchas"
Getting "make" to work on Windows isn't as easy as it should be.
You have to manually add the make binary directory to the PATH variable.
Also, (and I found this out the hard way) Powershell doesn't like the
Program Files (x86) directory so you'll have to install GnuMake32
somehwere else (e.g., C:\GnuMake32).  Ironic, isn't it?

Changes to be committed:
	modified:   README.md