KERI-14: Fixed hash_table solution collision BUG
It wasn't a problem with collision detection.
It was a logic error in _add_entry().
_add_entry() was calculating the actual index,
as opposed to collision detection, off the value instead of the key.

There's still some DEBUGGING statements (commented out) but it works,
with the current input, and ASAN is happy with it so I'm committing it.

Changes to be committed:
	modified:   src/8-04-hash_table-solution.c