KERI-14: Refactored example 8-04 hash table solut.
Refactored the example solution for load factor threshold support.
Now, the hash table will detect, at "add" time, when the load factor
exceeds the threshold and resize the hash table.

The legacy unit tests are still passing.
The new unit tests (see: Untracked below) are passing
ASAN and Valgrind are both happy.
I see this as an absolute win.

Changes to be committed:
	modified:   code/src/8-04-hash_table-solution.c

Changes not staged for commit:
	modified:   code/Makefile

Untracked files:
	code/src/8-04-3-hash_table-main.c