KERI-12: Implemented 8-02-2 sort_list() example
Implemented the sort_list() library API function.
I've been sitting on this uncommitted change for a couple days while I
tested it.  This version of the library appears to be passing all the
unit tests for valid reasons.  Also, ASAN and valgrind are happy.

It still needs to be cleaned up though.

Changes to be committed:
	modified:   code/src/8-02-linked_list-solution.c