KERI-12: 8-02-2 unit test/solution refactor EOD
EOD commit for the library refactor and unit test code (as it currently
is).

GOOD NEWS: The two unit tests currently implemented are passing.

BAD NEWS: The library solution is flush with DEBUGGING statements and
the unit test code has a huge swath of stale code.

EOD Commit...

Changes to be committed:
	new file:   code/src/8-02-2-circular_list-main.c
	modified:   code/src/8-02-2-circular_list-solution.c