KERI-15: Impromptu week 2 lab solution
These files represent a partial example solution.
void_lib.h is void_lib.txt turned into a header.
void_lib.c is a partial example solution from Tyler Walker that passes
all the unit tests, ASAN, and Valgrind in appropriate ways.

SPOILERS... All of this was done on-the-fly (see: live) in the
classroom.  I never got the chance to provide a complete
instructor-implementation of the library.

Changes to be committed:
	new file:   code/include/void_lib.h
	new file:   code/src/void_lib.c

Untracked files:
	code/src/23-2-C_Prog-W2-cap_test-1.c