KERI-13: 8-3 Binary Tree raw slides initial commit
I had started the work over a month ago but I think I added more to it
today than any of the previous days combined.

In my opinion, this is an interim commit of the (more detailed) binary
tree information.  BST content is yet to come...

Changes to be committed:
	new file:   slides_raw/08_03-C_Prog-Data_Structures-Binary_Tree.rst